Fixed Imperator currencies being ints instead of doubles (#125)
* Fixed Imperator currencies being ints instead of doubles

* Update googletest

* Revert "Update googletest"

This reverts commit 75db29abe29bd0d640e5ebb6c154567b53a2d350.

+++ ImperatorToCK3/Source/Imperator/Countries/Country.h
@@ -13,14 +13,14 @@ namespace Imperator
 {
 	typedef struct CurrenciesStruct
 	{
-		int manpower = 0;
-		int gold = 0;
-		int stability = 50;
-		int tyranny = 0;
-		int war_exhaustion = 0;
-		int aggressive_expansion = 0;
-		int political_influence = 0;
-		int military_experience = 0;
+		double manpower = 0;
+		double gold = 0;
+		double stability = 50;
+		double tyranny = 0;
+		double war_exhaustion = 0;
+		double aggressive_expansion = 0;
+		double political_influence = 0;
+		double military_experience = 0;
 	} CurrenciesStruct;
 
 	enum class countryTypeEnum { rebels, pirates, barbarians, mercenaries, real };
